# backend-flask/app.py
from flask import Flask, jsonify
from flask import request
from flask_cors import CORS, cross_origin
import os

from services.home_activities import *
from services.user_activities import *
from services.create_activity import *
from services.create_reply import *
from services.search_activities import *
from services.message_groups import *
from services.messages import *
from services.create_message import *
from services.show_activity import *

app = Flask(__name__)
frontend = os.getenv('FRONTEND_URL')
backend = os.getenv('BACKEND_URL')
origins = [frontend, backend]
cors = CORS(
    app,
    resources={r"/api/*": {"origins": origins}},
    expose_headers="location,link",
    allow_headers="content-type,if-modified-since",
    methods="OPTIONS,GET,HEAD,POST",
    allow_credentials=True
)  # added last line


@app.route('/health')
def health():
    return 'You are healthy!', 200


@app.route("/checkEnvVars", methods=['GET'])
def checkEnvVars():
    return frontend + " " + backend, 200


@app.route('/users')
def users():
    return jsonify({'users': ['Alice', 'Bob', 'Charlie']})


@app.route("/api/routes", methods=['GET'])
def routes():
    """This route shows all available routes"""
    routes = []
    for rule in app.url_map.iter_rules():
        route = {
            'url': str(rule),
            'methods': list(rule.methods),
            'endpoint': rule.endpoint
        }
        routes.append(route)
    data = jsonify(routes)
    return data, 200
    # data = jsonify({'routes': routes})   # react version
    # # return jsonify({'routes': routes}), 200  # json version
    # # data = HomeActivities.run()
    # return data, 200                     # react html version


@app.route("/api/message_groups", methods=['GET'])
def data_message_groups():
    user_handle = 'Jason'
    model = MessageGroups.run(user_handle=user_handle)
    if model['errors'] is not None:
        return model['errors'], 422
    else:
        return model['data'], 200


@app.route("/api/messages/@<string:handle>", methods=['GET'])
def data_messages(handle):
    user_sender_handle = 'Jason'
    user_receiver_handle = request.args.get('user_receiver_handle')

    model = Messages.run(user_sender_handle=user_sender_handle,
                         user_receiver_handle=user_receiver_handle)
    if model['errors'] is not None:
        return model['errors'], 422
    else:
        return model['data'], 200
    return


@app.route("/api/messages", methods=['POST', 'OPTIONS'])
@cross_origin()
def data_create_message():
    user_sender_handle = 'Jason'
    user_receiver_handle = request.json['user_receiver_handle']
    message = request.json['message']

    model = CreateMessage.run(
        message=message, user_sender_handle=user_sender_handle, user_receiver_handle=user_receiver_handle)
    if model['errors'] is not None:
        return model['errors'], 422
    else:
        return model['data'], 200
    return


@app.route("/api/activities/home", methods=['GET'])
def data_home():
    data = HomeActivities.run()
    return data, 200


@app.route("/api/activities/@<string:handle>", methods=['GET'])
def data_handle(handle):
    model = UserActivities.run(handle)
    if model['errors'] is not None:
        return model['errors'], 422
    else:
        return model['data'], 200


@app.route("/api/activities/search", methods=['GET'])
def data_search():
    term = request.args.get('term')
    model = SearchActivities.run(term)
    if model['errors'] is not None:
        return model['errors'], 422
    else:
        return model['data'], 200
    # return


@app.route("/api/activities", methods=['POST', 'OPTIONS'])
@cross_origin()
def data_activities():
    user_handle = 'Jason'
    message = request.json['message']
    ttl = request.json['ttl']
    model = CreateActivity.run(message, user_handle, ttl)
    if model['errors'] is not None:
        return model['errors'], 422
    else:
        return model['data'], 200
    # return


@app.route("/api/activities/<string:activity_uuid>", methods=['GET'])
def data_show_activity(activity_uuid):
    data = ShowActivity.run(activity_uuid=activity_uuid)
    return data, 200


@app.route("/api/activities/<string:activity_uuid>/reply", methods=['POST', 'OPTIONS'])
@cross_origin()
def data_activities_reply(activity_uuid):
    user_handle = 'Jason'
    message = request.json['message']
    model = CreateReply.run(message, user_handle, activity_uuid)
    if model['errors'] is not None:
        return model['errors'], 422
    else:
        return model['data'], 200
    # return


@app.route('/', defaults={'path': ''})
@app.route('/<path:path>')
def catch_all(path):
    """This route shows all available routes"""
    data = []
    warning = f"You entered '{path}'. This URL does not have a rule. Please return home. 🏘️🏡🐱🐈😹🐈‍⬛🐶🐕🦮🐕‍🦺 For your convenience we have provided the API routes that are available below as a reference guide."
    for rule in app.url_map.iter_rules():
        # route = {
        #     # 'url': str(rule),
        #     # 'methods': list(rule.methods),
        #     'endpoint': rule.endpoint
        # }
        # data.append(route)
        data.append(rule.endpoint)
    # json version
    return jsonify({'path': warning, 'routes': data}), 200

    # @app.route('/')
    # def home():
    #     # data = HomeActivities.run()
    #     # return data, 200
    #     return 'Welcome Home!🏠🏘️🏡🐱🐈😹🐈‍⬛🐶🐕🦮🐕‍🦺', 200


if __name__ == "__main__":
    app.run(debug=True)